requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0
pydantic-settings>=2.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.1
httpx>=0.27.0
fastapi-mail>=1.4.1
aiosmtplib>=3.0.0
remnawave>=2.0.0; python_version < "3.14"
pytest>=8.2.0
pytest-asyncio>=0.23.0
factory_boy>=3.3.0
aiosqlite>=0.20.0
